+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+-   Binary encoding and protocol.
+-   User Settings Bluetooth Service (USS).
+
 ## [1.2.0] - 2023-01-18
 
 ### Added